Add DifferOutputBuilderInterface.
